% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anc_desc_table.R
\name{anc_desc_table}
\alias{anc_desc_table}
\title{Get ancestor-descendant relations a given rc pair}
\usage{
anc_desc_table(
  i = 1,
  rc_table,
  rowTerminals,
  colTerminals,
  rowDescendants,
  colDescendants
)
}
\arguments{
\item{i}{integer: row of \code{rc_table} for extraction of ancestor and descendant \code{rc_index} relations}

\item{rc_table}{see \code{\link{make_rc_table}}}

\item{rowTerminals}{Indexes of terminal row tree edges}

\item{colTerminals}{Indexes of terminal column tree edges}

\item{rowDescendants}{List of descendants made from \code{edge_registry}. Element \code{i} contains all edges descendant from edge \code{i} in the row tree}

\item{colDescendants}{List of descendants made from \code{edge_registry}. Element \code{j} contains all edges descendant from edge \code{j} in the column tree}
}
\description{
Get ancestor-descendant relations a given rc pair
}
\examples{

# none yet - for internal dendromap use
}
