% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestorTable.R
\name{ancestorTable}
\alias{ancestorTable}
\title{get table of ancestral node-pair scores}
\usage{
ancestorTable(U, row.node, col.node, row.tree, col.tree)
}
\arguments{
\item{U}{score matrix}

\item{row.node}{node in \code{row.tree}}

\item{col.node}{node in \code{col.tree}}

\item{row.tree}{phylo class object whose i'th node (starting from the root being 1) corresponds to i'th row of \code{U}}

\item{col.tree}{phylo class object whose j'th node (starting from the root being 1) corresponds to j'th column of \code{U}}
}
\description{
get table of ancestral node-pair scores
}
