% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basal_dominance_filter.R
\name{basal_dominance_filter}
\alias{basal_dominance_filter}
\title{Remove all (r,c) such that P(r,c)>P(r',c) for some descendant r' of r}
\usage{
basal_dominance_filter(rc_tbl, rowDescendants)
}
\arguments{
\item{rc_tbl}{see \code{\link{make_rc_table}}}

\item{rowDescendants}{List of descendants made from \code{edge_registry}. Element \code{i} contains all edges descendant from edge \code{i} in the row tree}
}
\description{
Remove all (r,c) such that P(r,c)>P(r',c) for some descendant r' of r
}
