% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basalMax.R
\name{basalMax}
\alias{basalMax}
\title{Find maximum nodes in u along independent nodepaths to root}
\usage{
basalMax(u, row.tree, threshold = 0)
}
\arguments{
\item{u}{vector of length equal to \code{row.tree$Nnode}}

\item{row.tree}{\code{phylo} class object}

\item{threshold}{absolute value below which nodes and nodepaths are dropped from consideration. Default is 0}
}
\description{
Find maximum nodes in u along independent nodepaths to root
}
\examples{
set.seed(3)
m=1e3
n=30
row.tree <- rtree(m) \%>\% phytools::force.ultrametric()
col.tree <- rtree(n)

S <- treeSim(10,row.tree,col.tree,prob.row=0.7,prob.col=0.8,
             col.node = n+1,fix.col.node = T,sd = 1e3,
             row.depth.min=2,row.depth.max=3)

X <- S$W \%*\% S$D \%*\% t(S$V)
probs <- binomial(link='logit')$linkinv(X)
probs[X==0] <- 0.1
P <- rbinom(m*n,1,c(probs))
P <- matrix(P,nrow=m,ncol=n,byrow=F)

W <- treeBasis(row.tree)
V <- treeBasis(col.tree)
U <- t(W) \%*\% P \%*\% V

mx <- basalMax(U[,1],row.tree)
S$Path[row.node \%in\% mx$row.node]
}
