% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basalMaxima.R
\name{basalMaxima}
\alias{basalMaxima}
\title{Find maximum nodes in u along independent nodepaths to root}
\usage{
basalMaxima(X, row.tree, col.tree = NULL, W = NULL, V = NULL,
  threshold = 5, ncores = NULL)
}
\arguments{
\item{X}{Data matrix of number of rows equal to \code{length(row.tree$tip.label)} and number of columns equal to \code{length(col.tree$tip.label)}}

\item{row.tree}{\code{phylo} class object}

\item{col.tree}{\code{phylo} class object}

\item{W}{appropriately named \code{\link{treeBasis}} for \code{row.tree} or subset of that basis.}

\item{V}{appropriately named \code{\link{treeBasis}} for \code{col.tree} or subset of that basis.}

\item{threshold}{absolute value below which nodes and nodepaths are dropped from consideration. Default is 0}

\item{ncores}{Number of cores for parallel computation of \code{basalMax}}
}
\description{
Find maximum nodes in u along independent nodepaths to root
}
\examples{
set.seed(3)
m=1e3
n=30
row.tree <- rtree(m) \%>\% phytools::force.ultrametric()
col.tree <- rtree(n)

S <- treeSim(10,row.tree,col.tree,prob.row=0.7,prob.col=0.8,
             col.node = n+1,fix.col.node = T,sd = 1e3,
             row.depth.min=2,row.depth.max=3)

X <- makeDataset(S,family='gaussian')
Maxima <- basalMaxima(X,row.tree,col.tree,ncores=2)
Maxima[,orientation:=sign(value)]

### code copied from pathViz
colmap <- viridis::viridis(nrow(Maxima))
par(mfrow=c(1,1))
plot(row.tree,main='Row tree')
nodelabels(text=rep(' ',nrow(Maxima)),node=Maxima$row.node,bg = colmap,frame = 'circle')
nodelabels(text=rep('.',nrow(S$Paths)),node=S$Paths$row.node,bg='red',frame='circle',cex=.5)
}
