% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDescendants.R
\name{countDescendants}
\alias{countDescendants}
\title{Count descendants of a node}
\usage{
countDescendants(node, tree, direction = "pos")
}
\arguments{
\item{node}{numeric node of \code{tree} which has only two daughter clades}

\item{tree}{phylo class object}

\item{direction}{characer, either "pos" or "neg", indicating the direction of descendants to count (e.g. which daughter clade to count)/}
}
\description{
Count descendants of a node
}
