% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendroMap.R
\name{dendromap}
\alias{dendromap}
\title{dendromap - find cophylogenetic patterns in a dataset}
\usage{
dendromap(X, row.tree, col.tree, Pval_threshold = 0.01)
}
\arguments{
\item{X}{matrix whose rownames are in \code{row.tree$tip.label} and colnames are in \code{col.tree$tip.label}}

\item{row.tree}{phylo class object. Polytomies will be ignored.}

\item{col.tree}{phylo class object. Polytomies will be ignored.}

\item{Pval_threshold}{Threshold "significance" of a row-column node pair for consideration in lineages}
}
\description{
dendromap - find cophylogenetic patterns in a dataset
}
\examples{
library(dendromap)
set.seed(1)
m=1000
n=10
row.tree <- rtree(m)
col.tree <- rtree(n)
S <- treeSim(5,row.tree,col.tree,col.node=n+1)
eta <- S$W \%*\% (10*S$D) \%*\% t(S$V)
X <- eta+matrix(rnorm(m*n),nrow=m)
clrinv <- function(x) exp(x)/sum(exp(x))
rmlt <- function(p,lambda=5e3) rmultinom(1,rpois(1,lambda),prob = p)
N <- apply(X,2,clrinv) \%>\% apply(2,rmlt)
rownames(N) <- row.tree$tip.label
colnames(N) <- col.tree$tip.label
dm <- dendromap(N,row.tree,col.tree) 

plot.treesim(S)
plot.treesim(dm$Lineages)
S$Paths[,rc:=paste(row.node,col.node,sep='_')]
dm$Lineages[,rc:=paste(row.node,col.node,sep='_')]
sum(S$Paths$rc \%in\% dm$Lineages$rc)/nrow(S$Paths)      ### 64\% of the real rc's were ID'd
sum(dm$Lineages$rc \%in\% S$Paths$rc)/nrow(dm$Lineages)  ### at a 53\% FPR
}
