% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthDraw.R
\name{depthDraw}
\alias{depthDraw}
\title{draw node pairs for treeMap}
\usage{
depthDraw(d1, d2, lambda)
}
\arguments{
\item{d1}{depths on row tree}

\item{d2}{depths on col tree}

\item{lambda}{propensities of column tree nodes}
}
\description{
draw node pairs for treeMap
}
