% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disjointNodeset.R
\name{disjointNodeset}
\alias{disjointNodeset}
\title{Find nodes not on rootpaths between a set of nodes and their descendants}
\usage{
disjointNodeset(nodes, tree)
}
\arguments{
\item{nodes}{nodes of input \code{tree}}

\item{tree}{\code{phylo} class object}
}
\description{
Find nodes not on rootpaths between a set of nodes and their descendants
}
\examples{
set.seed(1)
library(ape)
tree <- rtree(10)
nodes <- c(15,19)
other.nodes <- disjointNodeset(nodes,tree)
plot(tree)
nodes <- setdiff(Ntip(tree)+1:tree$Nnode,other.nodes)
nodelabels(nodes,nodes,bg=rgb(0,.5,.5,1))
nodelabels(c(15,19),c(15,19),col='red')
nodelabels(other.nodes,other.nodes,bg=rgb(.9,.9,0,1))
}
