% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_ancestors.R
\name{edge_ancestors}
\alias{edge_ancestors}
\title{get path of ancestral edges all the way to, but not including, root}
\usage{
edge_ancestors(edge, tree)
}
\arguments{
\item{edge}{integer - which edge}

\item{tree}{\code{phylo} class object}
}
\description{
get path of ancestral edges all the way to, but not including, root
}
\examples{

}
