% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_contrasts.R
\name{edge_contrasts}
\alias{edge_contrasts}
\title{Create edge contrast vectors for tree}
\usage{
edge_contrasts(tree)
}
\arguments{
\item{tree}{phylo class object}
}
\description{
Create edge contrast vectors for tree
}
