% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_registry.R
\name{edge_registry}
\alias{edge_registry}
\title{Make registry of edges and number of descendants for quick traversals}
\usage{
edge_registry(tree)
}
\arguments{
\item{tree}{phylo class object}
}
\description{
Make registry of edges and number of descendants for quick traversals
}
