% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_tips.R
\name{edge_tips}
\alias{edge_tips}
\title{Table of edges and max/min values of descendant tips}
\usage{
edge_tips(tree)
}
\arguments{
\item{tree}{\code{phylo} class object}
}
\description{
Table of edges and max/min values of descendant tips
}
