% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_stats.R
\name{filter_stats}
\alias{filter_stats}
\title{Plot dendromap object}
\usage{
filter_stats(x, Lineages, rc_table, rc_relations, rowDescendants, row.tree)
}
\arguments{
\item{x}{output from \code{\link{lineage_stats}}}

\item{Lineages}{see \code{\link{get_lineages_and_stats}}}

\item{rc_table}{see \code{\link{make_rc_table}}}

\item{rc_relations}{see \code{\link{get_rc_relations}}}

\item{rowDescendants}{List of descendants made from \code{edge_registry}. Element \code{i} contains all edges descendant from edge \code{i} in the row tree}

\item{row.tree}{phylo class object}
}
\description{
Plot dendromap object
}
