% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_joinables.R
\name{find_joinables}
\alias{find_joinables}
\title{internal function to find joinable RC sequences}
\usage{
find_joinables(Seqs, rc_table, cl = NULL)
}
\arguments{
\item{Seqs}{output from \code{\link{rc_seqs}}}

\item{rc_table}{output from \code{\link{makeRCtable}}}

\item{cl}{optional cluster with dendromap loaded}
}
\description{
internal function to find joinable RC sequences
}
