% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLineageTable.R
\name{findLineageTable}
\alias{findLineageTable}
\title{Find table of lineages given rc_table subset by a P-value threshold}
\usage{
findLineageTable(rc_tbl, row.nodemap, col.nodemap, cl)
}
\arguments{
\item{rc_tbl}{table from \code{\link{makeRCtable}}}

\item{row.nodemap}{row tree nodemap from \code{\link{makeNodeMap}}}

\item{col.nodemap}{col tree nodemap from \code{\link{makeNodeMap}}}

\item{cl}{optional input cluster with package \code{dendromap} functions}
}
\description{
Find table of lineages given rc_table subset by a P-value threshold
}
