% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNodeSeq.R
\name{findNodeSeq}
\alias{findNodeSeq}
\title{find best sequence of node-pairs in the correct orientation as the column tree}
\usage{
findNodeSeq(RowCol, U. = U, row.tree. = row.tree,
  col.tree. = col.ree, row.nodemap. = row.nodemap,
  col.nodemap. = col.nodemap)
}
\arguments{
\item{RowCol}{vector of length two containing indexes \code{(i,j)} mapping the row tree node represented in row \code{i} of \code{U} and column tree node represented in column \code{j} of \code{U}}

\item{U}{matrix whose entry in row i column j is the bilinear map from row-tree's i'th node and the column tree's j'th node}

\item{row.tree}{phylo class object}

\item{col.tree}{phylo class object}

\item{row.nodemap}{nodemap produced in \code{\link{makeNodeMap}}}

\item{col.nodemap}{nodemap produced in \code{\link{makeNodeMap}}}
}
\description{
find best sequence of node-pairs in the correct orientation as the column tree
}
