% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineage_par.R
\name{get_lineage_par}
\alias{get_lineage_par}
\title{Build lineage from a given rc_index, using only P <= p_thresh}
\usage{
get_lineage_par(
  basal_ix,
  p_thresh = 1,
  rc_relations. = rc_relations,
  rc_table. = rc_table,
  row.tree. = row.tree,
  rowDescendants. = rowDescendants,
  colDescendants. = colDescendants
)
}
\arguments{
\item{basal_ix}{index found in \code{rc_table$rc_index}}

\item{p_thresh}{Positive numeric, threshold P-value for inclusion of rc-pairs}

\item{rc_relations}{}

\item{rc_table}{see \code{\link{make_rc_table}}}

\item{row.tree}{\code{phylo} class object}

\item{rowDescendants}{List of descendants made from \code{edge_registry}. Element \code{i} contains all edges descendant from edge \code{i} in the row tree}

\item{colDescendants}{List of descendants made from \code{edge_registry}. Element \code{j} contains all edges descendant from edge \code{j} in the column tree}
}
\description{
Build lineage from a given rc_index, using only P <= p_thresh
}
