% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phylo_groups.R
\name{get_phylo_groups}
\alias{get_phylo_groups}
\title{Get list of groups defined by phylogeny}
\usage{
get_phylo_groups(tree)
}
\arguments{
\item{tree}{Phylogeny}
}
\value{
list of groups and their complements, i.e. lists of groups split by each edge of the unrooted phylogeny
}
\description{
Get list of groups defined by phylogeny
}
\examples{
tr <- rtree(8)
tr$tip.label <- 1:8
par(mfrow=c(1,1))
plot.phylo(tr,use.edge.length=FALSE);edgelabels()
Grps <- get_phylo_groups(tr)
}
