% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rc_relations.R
\name{get_rc_relations}
\alias{get_rc_relations}
\title{Get all ancester-descendant relations of rc pairs}
\usage{
get_rc_relations(rc_table, rowDescendants, colDescendants)
}
\arguments{
\item{rc_table}{see \code{\link{make_rc_table}}}

\item{rowDescendants}{List of descendants made from \code{edge_registry}. Element \code{i} contains all edges descendant from edge \code{i} in the row tree}

\item{colDescendants}{List of descendants made from \code{edge_registry}. Element \code{j} contains all edges descendant from edge \code{j} in the column tree}
}
\description{
Get all ancester-descendant relations of rc pairs
}
