% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFstat.R
\name{getFstat}
\alias{getFstat}
\title{Calculate F-statistic for low-rank approximation X=WDV'+E for lineages in lineage_table}
\usage{
getFstat(X, lineage_table, W, V)
}
\arguments{
\item{X}{data matrix}

\item{lineage_table}{output from e.g. \code{\link{findLineageTable}}}

\item{W}{row tree-basis, \code{\link{treeBasis}}}

\item{V}{col tree-basis, \code{\link{treeBasis}}}
}
\description{
Calculate F-statistic for low-rank approximation X=WDV'+E for lineages in lineage_table
}
