% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndexSets.R
\name{getIndexSets}
\alias{getIndexSets}
\title{Get index sets of nodes descendant from a focal node, using nodemap}
\usage{
getIndexSets(nd, nodemap)
}
\arguments{
\item{nd}{node found in \code{nodemap$node}}

\item{nodemap}{data table produced in \code{\link{makeNodeMap}}}
}
\description{
Get index sets of nodes descendant from a focal node, using nodemap
}
