% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_Fstat.R
\name{global_Fstat}
\alias{global_Fstat}
\title{Extract global F statistic for several Lineages}
\usage{
global_Fstat(Lineages, X, colEdgeTips, rowEdgeTips)
}
\arguments{
\item{Lineages}{see \code{\link{get_lineages_and_stats}}}

\item{X}{dataset input to \code{\link{dendromap}} from which lineages were made}

\item{colEdgeTips}{\code{\link{edge_tips}} for col.tree}

\item{rowEdgeTips}{\code{\link{edge_tips}} for row.tree}
}
\description{
Extract global F statistic for several Lineages
}
