% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_lineage.R
\name{grow_lineage}
\alias{grow_lineage}
\title{grow_lineage from row/col edge pair. internal function for rdendromap}
\usage{
grow_lineage(
  row.edge,
  col.edge,
  row.tree,
  col.tree,
  rowEdgeMap,
  colEdgeMap,
  lineage_size,
  rowDescendants,
  colDescendants
)
}
\description{
grow_lineage from row/col edge pair. internal function for rdendromap
}
