% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idComparables.R
\name{idComparables}
\alias{idComparables}
\title{Identify rc pairs for which we have data to compare significance}
\usage{
idComparables(N, W, V, cl = NULL, verbose = TRUE)
}
\arguments{
\item{N}{dataset}

\item{W}{row.tree treeBasis}

\item{V}{col.tree treeBasis}

\item{cl}{cluster}

\item{verbose}{logical, whether or not to display message (used inside dendromap for more useful updates of algorithm progress)}
}
\description{
Identify rc pairs for which we have data to compare significance
}
\examples{

# If a row tree has descendant clades represented by {clade1,clade2} and col tree has descendant clades {continent1,continent2}
# Then a row-col node pair is considered 'incomparable' if any row or column of its contingency table has all zeros.
# e.g.: incomparable:
#       cnt1   cnt2  |        cnt1   cnt2
# spp1   0      0    |  spp1   0     0.5
# spp2   1     0.5   |  spp2   0      1
#
#e.g. comparable:
#
#       cnt1   cnt2  |        cnt1   cnt2
# spp1   0      1    |  spp1   0     0.5
# spp2   1     0.5   |  spp2   1      1
}
