% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incompatible_descendants.R
\name{incompatible_descendants}
\alias{incompatible_descendants}
\title{Find descendant rc_indexes incompatible with input rc_index}
\usage{
incompatible_descendants(
  rc_ix,
  rc_tbl,
  rowDescendants,
  colDescendants,
  row.tree. = row.tree,
  col.tree. = col.tree
)
}
\arguments{
\item{rc_ix}{index found in \code{rc_tbl$rc_index}}

\item{rc_tbl}{see \code{\link{make_rc_table}}. Internally in \code{\link{dendromap}}, this is a trimmed version of the global \code{rc_table}}

\item{rowDescendants}{List of descendants made from \code{edge_registry}. Element \code{i} contains all edges descendant from edge \code{i} in the row tree}

\item{colDescendants}{List of descendants made from \code{edge_registry}. Element \code{j} contains all edges descendant from edge \code{j} in the column tree}

\item{row.tree}{\code{phylo} class object}

\item{col.tree}{\code{phylo} class object}
}
\description{
Find descendant rc_indexes incompatible with input rc_index
}
