% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelledChildren.R
\name{labelledChildren}
\alias{labelledChildren}
\title{label children of a node to keep track of tree orientation}
\usage{
labelledChildren(node, tree)
}
\arguments{
\item{node}{internal node of \code{tree}}

\item{tree}{\code{phylo} class object}
}
\description{
label children of a node to keep track of tree orientation
}
\examples{
set.seed(1)
library(ape)
tree <- rtree(10)
labelledChildren(11,tree)
}
