% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_boxes.R
\name{lineage_boxes}
\alias{lineage_boxes}
\title{Coarse-grain dataset, X, based on discovered lineage}
\usage{
lineage_boxes(lineage, X, colEdgeTips, rowEdgeTips)
}
\arguments{
\item{lineage}{see \code{\link{get_lineage}}}

\item{X}{dataset input to \code{\link{dendromap}} from which lineages were made}

\item{colEdgeTips}{\code{\link{edge_tips}} for col.tree}

\item{rowEdgeTips}{\code{\link{edge_tips}} for row.tree}
}
\description{
Coarse-grain dataset, X, based on discovered lineage
}
