% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_plot.R
\name{lineage_plot}
\alias{lineage_plot}
\title{Focused of particular lineage}
\usage{
lineage_plot(dm, lineage = NULL, id = NULL, ...)
}
\arguments{
\item{dm}{dendromap object}

\item{lineage}{lineage number in \code{dm$Lineage} for plotting}

\item{id}{Alternative to \code{lineage}, \code{dm$lineage_id} for plotting}

\item{...}{additional input arguments for \code{\link{plot.dendromap}}}
}
\description{
Focused of particular lineage
}
