% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_stats.R
\name{lineage_stats}
\alias{lineage_stats}
\title{Extract F statistics for all lineages}
\usage{
lineage_stats(Lineages, X, colEdgeTips, rowEdgeTips, cl = NULL)
}
\arguments{
\item{Lineages}{see \code{\link{get_lineage}}}

\item{X}{dataset input to \code{\link{dendromap}} from which lineages were made}

\item{colEdgeTips}{\code{\link{edge_tips}} for col.tree}

\item{rowEdgeTips}{\code{\link{edge_tips}} for row.tree}

\item{cl}{cluster initialized internally with \code{\link{dendromap}}}
}
\description{
Extract F statistics for all lineages
}
