% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstat.R
\name{lstat}
\alias{lstat}
\title{Extract F statistic for a given lineage}
\usage{
lstat(
  id,
  Lineages,
  X. = X,
  colEdgeTips. = colEdgeTips,
  rowEdgeTips. = rowEdgeTips
)
}
\arguments{
\item{id}{value found in \code{Lineages$lineage_id}}

\item{Lineages}{see \code{\link{get_lineage}}}

\item{X}{dataset input to \code{\link{dendromap}} from which lineages were made}

\item{colEdgeTips}{\code{\link{edge_tips}} for col.tree}

\item{rowEdgeTips}{\code{\link{edge_tips}} for row.tree}
}
\description{
Extract F statistic for a given lineage
}
