% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rc_table.R
\name{make_rc_table}
\alias{make_rc_table}
\title{Make table of {row tree,column tree} edge pair statistics}
\usage{
make_rc_table(X, row.tree, col.tree, maxPval = 0.01)
}
\arguments{
\item{X}{dataset whose rows correspond exactly to \code{row.tree$tip.label} and columns correspond to \code{col.tree$tip.label}. NA values not allowed.}

\item{row.tree}{\code{phylo} class object}

\item{col.tree}{\code{phylo} class object}

\item{maxPval}{Cutoff - remove all P-values greater than \code{maxPval}}
}
\description{
Make table of {row tree,column tree} edge pair statistics
}
\examples{
## none yet
}
