% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataset.R
\name{makeDataset}
\alias{makeDataset}
\title{make dataset from treemap object}
\usage{
makeDataset(Path, row.tree, col.tree, d = NULL, sd = 10,
  baseline = 0, family = "bernoulli", ...)
}
\arguments{
\item{Path}{output from \code{\link{treeMap}}}

\item{row.tree}{\code{phylo} class object used to make \code{Path}}

\item{col.tree}{\code{phylo} class object used to make \code{Path}}

\item{d}{numeric vector of length equal to \code{nrow(Path)}}

\item{sd}{standard deviation for simulation d, if d is not input.}

\item{baseline}{baseline rate parameter for random variables not given a value from Path}

\item{family}{string, either "bernoulli" or "poisson" is supported at this time, indicating the family of random variables to simulate in dataset}
}
\description{
make dataset from treemap object
}
