% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDescendantTable.R
\name{makeDescendantTable}
\alias{makeDescendantTable}
\title{Internal functin helping makeRCmap find rc-rc ancestor-descendant relations}
\usage{
makeDescendantTable(i = 1, rc_table, terminalRowNodes, terminalColNodes,
  Row_Descendants, Col_Descendants)
}
\arguments{
\item{i}{index of rc_table whose descendants are found}

\item{terminalRowNodes}{row nodes in rc_table with no descendants}

\item{terminalColNodes}{same for col nodes in col tree}

\item{Row_Descendants}{list of descendants for each row node}

\item{Col_Descendants}{list of descendants for each col node}
}
\description{
Internal functin helping makeRCmap find rc-rc ancestor-descendant relations
}
