% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNodeMap.R
\name{makeNodeMap}
\alias{makeNodeMap}
\title{make nodemap for quick referencing of descendants}
\usage{
makeNodeMap(tree)
}
\arguments{
\item{tree}{phylo class object}
}
\description{
make nodemap for quick referencing of descendants
}
