% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRCmap.R
\name{makeRCMap}
\alias{makeRCMap}
\title{make rc-rc mapping for ancestor-descendant relationships & rc tree traversal}
\usage{
makeRCMap(rc_table, Row_Descendants, Col_Descendants)
}
\arguments{
\item{rc_table}{made from \code{\link{makeRCtable}}}

\item{Row_Descendants}{named \code{getIndexSets} of all row.nodes}

\item{Col_Descendants}{named \code{getIndexSets} of all col.nodes}
}
\description{
make rc-rc mapping for ancestor-descendant relationships & rc tree traversal
}
