% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRCtable.R
\name{makeRCtable}
\alias{makeRCtable}
\title{make table of significant row-column node pairs}
\usage{
makeRCtable(N, row.tree, col.tree, W = NULL, V = NULL, n_sim = NULL)
}
\arguments{
\item{N}{matrix whose row names are in \code{row.tree$tip.label} and column names are in \code{col.tree$tip.label}}

\item{row.tree}{phylo class object}

\item{col.tree}{phylo class object}

\item{W}{\code(treeBasis(row.tree))}

\item{V}{\code(treeBasis(col.tree))}

\item{n_sim}{number of simulatd null datasets (used to compuate approximate P values for r-c node pairs)}
}
\description{
make table of significant row-column node pairs
}
