% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRowColSet.R
\name{makeRowColSet}
\alias{makeRowColSet}
\title{make RowCol list for parallelized input to findNodeSeq}
\usage{
makeRowColSet(U, forbidden.nodes = NULL)
}
\arguments{
\item{U}{matrix of bilinear mappings from {row.tree}x{col.tree} nodes}

\item{forbidden.nodes}{integer vector of row nodes to exclude from consideration}
}
\description{
make RowCol list for parallelized input to findNodeSeq
}
