% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifestChildren.R
\name{manifestChildren}
\alias{manifestChildren}
\title{Find all manifested descendants of a node}
\usage{
manifestChildren(node, tree, nodebank)
}
\arguments{
\item{node}{node in \code{tree}}

\item{tree}{\code{phylo} class object}

\item{nodebank}{output from \code{\link{nodeBank}}}
}
\description{
Find all manifested descendants of a node
}
\examples{
set.seed(1)
library(ape)
tree <- rtree(10)
nb <- nodeBank(11,tree,0.8)
manifestChildren(11,tree,nb)
}
