% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifestRowChildren.R
\name{manifestRowChildren}
\alias{manifestRowChildren}
\title{Manifest children for row tree, sensitive to depth of col.nodes}
\usage{
manifestRowChildren(node, row.tree, row.nb, col.nb, col.nodes,
  use.depths = F)
}
\arguments{
\item{node}{row tree node whose children need to be found}

\item{row.tree}{\code{phylo} class object}

\item{row.nb}{\code{\link{nodeBank}} for \code{row.tree}}

\item{col.nb}{\code{\link{nodeBank}} for \code{col.tree}}

\item{col.nodes}{nodes from column tree}

\item{use.depths}{logical - whether or not to use depth-based filtering through propensities in \code{col.nb}}
}
\description{
Manifest children for row tree, sensitive to depth of col.nodes
}
