% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchAncestors.R
\name{matchAncestors}
\alias{matchAncestors}
\title{Get scores for ancestor sequences}
\usage{
matchAncestors(U, row.node, col.node, row.tree, col.tree,
  orientation.match = TRUE)
}
\arguments{
\item{U}{score matrix whose i'th row corresponds to the (starting from root=1) i'th node in \code{row.tree} and j'th column corresponds to the j'th node in \code{col.tree}}

\item{row.node}{node in \code{row.tree}}

\item{col.node}{node in \code{col.tree}}

\item{row.tree}{phylo class object}

\item{col.tree}{phylo class object}

\item{orientation.match}{Logical. Whether orientation of paths must be perfect match. Currently only supports \code{TRUE} until soft-matching is coded up.}
}
\description{
Get scores for ancestor sequences
}
\examples{
library(dendromap)
set.seed(1)
row.tree <- ape::rtree(50)
col.tree <- ape::rtree(10)
U <- t(treeBasis(row.tree)) \%*\% matrix(rnorm(500),nrow=50) \%*\% treeBasis(col.tree)
par(mfrow=c(1,2))
plot(row.tree)
nodelabels()
nodelabels(77,77,bg='red')
plot(col.tree)
nodelabels()
nodelabels(16,16,bg='red')

Anc <- matchAncestors(U,77,16,row.tree,col.tree)
}
