% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDescendants.R
\name{matchDescendants}
\alias{matchDescendants}
\title{Match descendants using basalMaxima}
\usage{
matchDescendants(U, Seq, row.tree, col.tree, row.nodemap, col.nodemap,
  quantile.threshold = 0.95)
}
\description{
Match descendants using basalMaxima
}
