% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_twin_sisters.R
\name{merge_twin_sisters}
\alias{merge_twin_sisters}
\title{Identify sister row edges with same column edge association & merge at mother}
\usage{
merge_twin_sisters(lineage, row.tree, rc_table)
}
\arguments{
\item{lineage}{lineage table. See \code{\link{get_lineage}}}

\item{row.tree}{\code{phylo} class object}

\item{col.tree}{\code{phylo} class object}
}
\description{
Identify sister row edges with same column edge association & merge at mother
}
