% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeBank.R
\name{nodeBank}
\alias{nodeBank}
\title{Make a dataset of nodes in a tree and whether/not they are manifested}
\usage{
nodeBank(node, tree, prob = 1)
}
\arguments{
\item{node}{node in tree}

\item{tree}{\code{phylo} class object}

\item{prob}{probability between 0 and 1}
}
\description{
Make a dataset of nodes in a tree and whether/not they are manifested
}
\examples{
set.seed(1)
library(ape)
tree <- rtree(10)
nb <- NodeBank(11,tree,0.5)
}
