% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathViz.R
\name{pathViz}
\alias{pathViz}
\title{Visualize a single, manifested path on two trees}
\usage{
pathViz(Path, col.tree, row.tree)
}
\arguments{
\item{Path}{output from \code{\link{treeMap}}}

\item{col.tree}{\code{phylo} class object from \code{treeMap}}

\item{row.tree}{\code{phylo} class object from \code{treeMap}}
}
\description{
Visualize a single, manifested path on two trees
}
\examples{
set.seed(1)
library(ape)
row.tree <- rtree(100)
col.tree <- rtree(10)
S <- treeMap(row.tree,col.tree,101,11,0.2,1)
pathViz(S,col.tree,row.tree)
}
