% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dendromap.R
\name{plot.dendromap}
\alias{plot.dendromap}
\title{Plot dendromap object}
\usage{
\method{plot}{dendromap}(x, y = NULL, orient.nodes = TRUE,
  color.fcn.clade = viridis::viridis,
  color.fcn.node = viridis::viridis, heatmap.offset = 0,
  col.tr.left = 0.5, col.tr.width = 0.45, col.tr.bottom = 0.75)
}
\arguments{
\item{x}{dendromap class object from \code{\link{treeSim}} or \code{\link{dendromap}}}

\item{y}{optional data matrix. Must have labelled rows containing all tips in \code{x$row.tree} and likewise for \code{x$col.tree}}

\item{orient.nodes}{logical - whether/not to rotate all nodes to consistent orientation}

\item{color.fcn.clade}{color function for clade highlights in \code{\link{ggtree}}}

\item{color.fcn.node}{color function for node circles}

\item{heatmap.offset}{offset put into \code{\link{gheatmap}}. Defualt is 0}

\item{col.tr.left}{Left side boundary of column tree. Default is 0.5}

\item{col.tr.width}{Width of column tree. Default is 0.45}

\item{col.tr.bottom}{Bottom of column tree. Deafult is 0.75}
}
\description{
Plot dendromap object
}
\examples{
set.seed(3)
m=1e3
n=7
row.tree <- rtree(m) \%>\% phytools::force.ultrametric()
col.tree <- rtree(n)

S <- treeSim(10,row.tree,col.tree,prob.row=0.7,prob.col=0.8,
             col.node = n+1,fix.col.node = T,sd = 1e3,
             row.depth.min=2,row.depth.max=3)

plot(S,col.tr.left = 0.47,
     col.tr.width = 0.505,
     col.tr.bottom = 0.74)
}
