% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.dendromap.R
\name{predict.dendromap}
\alias{predict.dendromap}
\title{Predict dataset from dendromap object}
\usage{
\method{predict}{dendromap}(object, ...)
}
\arguments{
\item{object}{\code{\link{dendromap}} object}

\item{...}{additional arguments, currently none are used}
}
\description{
Predict dataset from dendromap object
}
