% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dendromap.R
\name{print.dendromap}
\alias{print.dendromap}
\title{print dendromap object}
\usage{
\method{print}{dendromap}(x)
}
\arguments{
\item{x}{\code{\link{dendromap}} object}
}
\description{
print dendromap object
}
