% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdendromap.R
\name{rdendromap}
\alias{rdendromap}
\title{Random dendromap object}
\usage{
rdendromap(
  m = NULL,
  n = NULL,
  row.tree = NULL,
  col.tree = NULL,
  lineage_sizes = NULL,
  lambda = 3,
  n_lineages = 1,
  min.lineage.size = 2,
  simulate.dataset = FALSE,
  ...
)
}
\arguments{
\item{m}{number of tips in \code{row.tree}}

\item{n}{number of tips in \code{col.tree}}

\item{row.tree}{optional phylo class input, instead of m.}

\item{col.tree}{optional phylo class input, instead of n.}

\item{lineage_sizes}{either single integer or integer vector of length \code{n_lineages}. Lineages will be less than or equal to \code{lineage_sizes}}

\item{lambda}{positive numeric rate parameter for \code{\link{rpois}} generating lineage sizes}

\item{n_lineages}{integer number of n_lineages to compute}

\item{min.lineage.size}{minimum lineage size -}
}
\description{
Random dendromap object
}
