% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootpath.R
\name{rootpath}
\alias{rootpath}
\title{find path from node to root}
\usage{
rootpath(node, tree)
}
\arguments{
\item{node}{node of \code{tree}}

\item{tree}{phylo class object}
}
\description{
find path from node to root
}
\examples{
library(dendromap)
set.seed(1)
tree <- rtree(20)
rp=rootpath(33,tree)

plot(tree)
nodelabels(rp,rp)
}
