% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanFstats.R
\name{scanFstats}
\alias{scanFstats}
\title{Scan a set of P-value cutoffs, returning F statistics for each}
\usage{
scanFstats(Pset, RC_table, row.nodemap, col.nodemap, W, V, cl)
}
\description{
Scan a set of P-value cutoffs, returning F statistics for each
}
