% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_rc_table.R
\name{simplify_rc_table}
\alias{simplify_rc_table}
\title{Remove (r,c) pairs without ancestors/descendants, and impose basal dominance}
\usage{
simplify_rc_table(rc_table, rc_relations, rowDescendants, colDescendants)
}
\arguments{
\item{rc_table}{see \code{\link{make_rc_table}}}

\item{rc_relations}{see \code{\link{get_rc_relations}}}

\item{rowDescendants}{List of descendants made from \code{edge_registry}. Element \code{i} contains all edges descendant from edge \code{i} in the row tree}

\item{colDescendants}{List of descendants made from \code{edge_registry}. Element \code{j} contains all edges descendant from edge \code{j} in the column tree}
}
\description{
Remove (r,c) pairs without ancestors/descendants, and impose basal dominance
}
