% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeBasis.R
\name{treeBasis}
\alias{treeBasis}
\title{Make basis from phylogenetic tree}
\usage{
treeBasis(tree)
}
\arguments{
\item{tree}{\code{phylo} class object}
}
\description{
Make basis from phylogenetic tree
}
\examples{
library(ape)
tree <- rtree(5)
treeBasis(tree)
}
