% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap.R
\name{treeMap}
\alias{treeMap}
\title{Map events on column tree to nodes on row tree}
\usage{
treeMap(row.tree, col.tree, row.node = NULL, col.node = NULL,
  prob.row = 1, prob.col = 1)
}
\arguments{
\item{row.tree}{\code{phylo} class object}

\item{col.tree}{\code{phylo} class object}

\item{row.node}{integer. optional node for starting treeMap. Must be a node in \code{row.tree}}

\item{col.node}{integer. optional node for starting treeMap. Must be a node in \code{col.tree}}

\item{prob.row}{probability, between 0 and 1, that a descendant node is manifested in the tree map.}

\item{prob.col}{same as \code{prob.row} but for nodes in column tree.}
}
\description{
Map events on column tree to nodes on row tree
}
\examples{
set.seed(1)
library(ape)
row.tree <- rtree(100)
col.tree <- rtree(10)
S <- treeMap(row.tree,col.tree,101,11,0.2,1)
pathViz(S,col.tree,row.tree)
}
