% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSim.R
\name{treeSim}
\alias{treeSim}
\title{Simulate multiple treeMaps on row.tree}
\usage{
treeSim(n, row.tree, col.tree, prob.row = 1, prob.col = 1, sd = 1,
  W = NULL, V = NULL, row.node = NULL, col.node = NULL,
  row.nodeset = NULL, col.nodeset = NULL, row.depth.min = NULL,
  row.depth.max = NULL, col.depth.min = NULL, col.depth.max = NULL,
  fix.col.node = FALSE, decay.rate = 0, use.depths = F,
  col.nb = NULL, row.nb = NULL)
}
\arguments{
\item{n}{max number of independent lineages on \code{row.tree} to map to column tree}

\item{row.tree}{\code{phylo} class object}

\item{col.tree}{\code{phylo} class object}

\item{prob.row}{probability of a node in the row.tree manifesting a map to col.tree. Default is 1}

\item{prob.col}{probability of a node in the col.tree manifesting a map to row.tree. Default is 1}

\item{sd}{optional positive numeric, standard deviation input to \code{\link{rnorm}} to create output diagonal matrix}

\item{W}{optional matrix, \code{\link{treeBasis}} of \code{row.tree}}

\item{V}{optional matrix, \code{\link{treeBasis}} of \code{col.tree}}

\item{row.node}{optional start node for row.tree}

\item{col.node}{optional start node for col.tree}

\item{row.nodeset}{optional set of nodes to consider for random start of \code{treeMap}}

\item{col.nodeset}{optional set of nodes to consider for random start of \code{treeMap}}

\item{row.depth.min}{optional numeric minimum depth for nodes for random start of \code{treeMap}}

\item{row.depth.max}{optional numeric maximum depth for nodes for random start of \code{treeMap}}

\item{col.depth.min}{optional numeric minimum depth for nodes for random start of \code{treeMap}}

\item{col.depth.max}{optional numeric maximum depth for nodes for random start of \code{treeMap}}

\item{fix.col.node}{logical: whether or not to fix the column node for the start of every \code{treeMap}}

\item{decay.rate}{exponential rate of decay for past nodes. Probability of draw will be exp(-decay.rate*dT) where dT is the time from column node to row node. Only implemented if \code{use.depths=TRUE}}

\item{col.nb}{\code{\link{nodeBank}} for column tree, allowing user to set up propensities of column tree nodes.}

\item{use.depth}{logical: whether or not to use propensities to limit manifestation in row tree to being after nodes in column tree.}
}
\description{
Simulate multiple treeMaps on row.tree
}
\examples{
library(ggplot2)
library(ggpubr)
set.seed(3)
m=1e3
n=30
row.tree <- rtree(m) \%>\% phytools::force.ultrametric()
col.tree <- rtree(n)
S <- treeSim(10,row.tree,col.tree,prob.row=0.7,prob.col=0.8,sd = 1e3,
             row.depth.min=2,row.depth.max=3)              

plot(S,col.tr.left = 0.47,
     col.tr.width = 0.505,
     col.tr.bottom = 0.74)
 
set.seed(1)
col.nb <- nodeBank(node=n+1,col.tree,propensity=50,include.node=T)
col.nb[,propensity:=50*(max(depth)-depth)]  ##root nodes have higher propensities
S <- treeSim(50,row.tree,col.tree,prob.row=0.7,prob.col=0.8,sd = 1e3,
             col.nb=col.nb,decay.rate=20,use.depths=T) 
             
dendroPlot <- plot(S,col.tr.left = 0.47,
                   col.tr.width = 0.505,
                   col.tr.bottom = 0.74)   
     
row.depths <- data.table('row.depth'=node.depth.edgelength(row.tree))
row.depths[,row.node:=1:.N]
setkey(row.depths,row.node)
col.depths <- data.table('col.depth'=node.depth.edgelength(col.tree))
col.depths[,col.node:=1:.N]
setkey(col.depths,col.node)
setkey(S$Paths,row.node)
Paths <- row.depths[S$Paths]
setkey(Paths,col.node)
Paths <- col.depths[Paths]
Paths[,Lineage:=factor(sim)]

depthPlot=ggplot(Paths,aes(row.depth,col.depth,color=Lineage))+
                 geom_point()+
                 geom_abline(intercept=0,slope=1)+
                 geom_smooth(method='glm',se=F)+
                 scale_x_continuous('Time on Microbial Tree')+
                 scale_y_continuous('Time on Reference Tree')
          
ggarrange(dendroPlot,depthPlot)
}
