% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warpTree.R
\name{warpTree}
\alias{warpTree}
\title{Warps tree}
\usage{
warpTree(tree, trans = "log", alpha = NULL)
}
\arguments{
\item{tree}{phylo class object}

\item{trans}{character, either 'log' or 'power', indicating whether to use \code{a*log(x/a+1)} or \code{x^a/a} warping}

\item{alpha}{see functional forms for \code{trans}. Default for 'log' is 1, default for 'power' is 1/2}
}
\description{
Warps tree
}
\examples{
set.seed(1)
tr <- rtree(100)
tr <- phytools::force.ultrametric(tr)
par(mfrow=c(2,1))
plot(tr,show.tip.label = F)
plot(warp.phylo(tr),show.tip.label = F)
}
