% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basalMaxima.R
\name{basalMaxima}
\alias{basalMaxima}
\title{Find maximum nodes in u along independent nodepaths to root}
\usage{
basalMaxima(X, row.tree, col.tree = NULL, W = NULL, V = NULL,
  threshold = NULL, min.val = -10, ncores = NULL)
}
\arguments{
\item{X}{Data matrix of number of rows equal to \code{length(row.tree$tip.label)} and number of columns equal to \code{length(col.tree$tip.label)}}

\item{row.tree}{\code{phylo} class object}

\item{col.tree}{\code{phylo} class object}

\item{W}{appropriately named \code{\link{treeBasis}} for \code{row.tree} or subset of that basis.}

\item{V}{appropriately named \code{\link{treeBasis}} for \code{col.tree} or subset of that basis.}

\item{threshold}{absolute value below which nodes and nodepaths are dropped from consideration. Default is \code{4*sd(log(U^2))}}

\item{min.val}{minimum value for \eqn{log(U^2)}, where \eqn{U=W'XV}. All values below \code{min.val} will be set to 0}

\item{ncores}{Number of cores for parallel computation of \code{basalMax}}
}
\description{
Find maximum nodes in u along independent nodepaths to root
}
\examples{
set.seed(3)
m=1e2
n=5
row.tree <- rtree(m) \%>\% phytools::force.ultrametric()
col.tree <- rtree(n)

S <- treeSim(10,row.tree,col.tree,prob.row=0.8,prob.col=0.8,
             col.node = n+1,fix.col.node = T,sd = 1e3,
             row.depth.min=2,row.depth.max=3)

X <- makeDataset(S,family='gaussian')
Maxima <- basalMaxima(X,row.tree,col.tree,ncores=2)
Maxima[,orientation:=sign(value)]

### code copied from pathViz
colmap <- data.table('color'=viridis::viridis(length(unique(Maxima$col.node))),
                     'node'=unique(Maxima$col.node))

par(mfrow=c(1,3))
plot(row.tree,main='Row tree, raw Gaussian')
nodelabels(text=rep(' ',nrow(Maxima)),node=Maxima$row.node,cex=2,
           bg = colmap[match(Maxima$col.node,node),color],frame = 'circle')
nodelabels(text=rep('.',nrow(S$Paths)),node=S$Paths$row.node,bg='red',frame='circle',cex=.5)

probs <- binomial(link='logit')$linkinv(X)
probs[X==0] <- 0.1
P <- rbinom(m*n,1,c(probs)) \%>\% matrix(nrow=m,byrow=F)
MaximaP <- basalMaxima(P,row.tree,col.tree,threshold=1.5)
plot(row.tree,main='Row tree, Bernoulli data, threshold=1.5')
nodelabels(text=rep(' ',nrow(MaximaP)),node=MaximaP$row.node,cex=2,
           bg = colmap[match(MaximaP$col.node,node),color],frame = 'circle')
nodelabels(text=rep('.',nrow(S$Paths)),node=S$Paths$row.node,bg='red',frame='circle',cex=.5)

MaximaP2 <- basalMaxima(P,row.tree,col.tree,threshold=0.1)
plot(row.tree,main='Row tree, Bernoulli data, threshold=0.1')
nodelabels(text=rep(' ',nrow(MaximaP2)),node=MaximaP2$row.node,cex=2,
           bg = colmap[match(MaximaP2$col.node,node),color],frame = 'circle')
nodelabels(text=rep('.',nrow(S$Paths)),node=S$Paths$row.node,bg='red',frame='circle',cex=.5)
}
