% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_joinable.R
\name{check_joinable}
\alias{check_joinable}
\title{check if two rc_seqs are joinable}
\usage{
check_joinable(seq1, seq2, Seqs, rc_table, Row_Descendants,
  Col_Descendants)
}
\arguments{
\item{seq1}{index of sequence from \code{\link{rc_seqs}}}

\item{seq2}{index of sequence from \code{\link{rc_seqs}}}

\item{Seqs}{Set of sequences from \code{\link{rc_seqs}}}

\item{rc_table}{made in \code{\link{makeRCtable}}}

\item{Row_Descendants}{named \code{getIndexSets} of all row.nodes}

\item{Col_Descendants}{named \code{getIndexSets} of all col.nodes}
}
\description{
check if two rc_seqs are joinable
}
