% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_joinable.R
\name{check_joinable}
\alias{check_joinable}
\title{check if two rc_seqs are joinable}
\usage{
check_joinable(seq1, seq2, Seqs, rc_table, r.nodemap, c.nodemap)
}
\arguments{
\item{seq1}{index of sequence from \code{\link{rc_seqs}}}

\item{seq2}{index of sequence from \code{\link{rc_seqs}}}

\item{Seqs}{Set of sequences from \code{\link{rc_seqs}}}

\item{rc_table}{made in \code{\link{makeRCtable}}}

\item{r.nodemap}{\code{makeNodeMap} of row.tree}

\item{c.nodemap}{\code{makeNodeMap} of col.tree}
}
\description{
check if two rc_seqs are joinable
}
