% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendromap.R
\name{dendromap}
\alias{dendromap}
\title{Coarse-grain dataset, X, based on discovered lineage}
\usage{
dendromap(X, row.tree, col.tree, ncores = NULL, maxPval = 0.001, stepsize = 10)
}
\arguments{
\item{X}{dataset input to \code{\link{dendromap}} from which lineages were made}

\item{row.tree}{\code{phylo} class object}

\item{col.tree}{\code{phylo} class object}

\item{ncores}{integer number of cores for parallelization}

\item{maxPval}{numeric, maximum P-value of \code{row.tree}-x-\code{col.tree} edge pairs to consider}

\item{stepsize}{granularity of pvals initially input to \code{\link{scan_Fstats}}}
}
\description{
Coarse-grain dataset, X, based on discovered lineage
}
