% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendroMap.R
\name{dendromap}
\alias{dendromap}
\title{dendromap - find cophylogenetic patterns in a dataset}
\usage{
dendromap(X, row.tree, col.tree, Pval_threshold = 0.01, W = NULL,
  V = NULL, n_sim = NULL)
}
\arguments{
\item{X}{matrix whose rownames are in \code{row.tree$tip.label} and colnames are in \code{col.tree$tip.label}}

\item{row.tree}{phylo class object. Polytomies will be ignored.}

\item{col.tree}{phylo class object. Polytomies will be ignored.}

\item{Pval_threshold}{Threshold "significance" of a row-column node pair for consideration in lineages}

\item{W}{optional \code{treeBasis(row.tree)} - must have colnames in the format of e.g. "node_51" for node 51.}

\item{V}{optional \code{treeBasis(col.tree)} - must have colnames in the format of e.g. "node_51" for node 51.}

\item{n_sim}{optional number of null datasets to simulate (via row & column shuffling) in order to generate approximate P-values in \code{\link{makeRCtable}}}
}
\description{
dendromap - find cophylogenetic patterns in a dataset
}
\examples{
library(dendromap)
set.seed(1)
m=1000
n=10
row.tree <- rtree(m)
col.tree <- rtree(n)
S <- treeSim(5,row.tree,col.tree,col.node=n+1)
eta <- S$W \%*\% (10*S$D) \%*\% t(S$V)
X <- eta+matrix(rnorm(m*n),nrow=m)
clrinv <- function(x) exp(x)/sum(exp(x))
rmlt <- function(p,lambda=5e3) rmultinom(1,rpois(1,lambda),prob = p)
N <- apply(X,2,clrinv) \%>\% apply(2,rmlt)
rownames(N) <- row.tree$tip.label
colnames(N) <- col.tree$tip.label
dm <- dendromap(N,row.tree,col.tree) 

plot.dendromap(S)
plot.treesim(dm$Lineages)
S$Lineages[,rc:=paste(row.node,col.node,sep='_')]
dm$Lineages[,rc:=paste(row.node,col.node,sep='_')]
sum(S$Lineages$rc \%in\% dm$Lineages$rc)/nrow(S$Lineages)      ### 64\% of the real rc's were ID'd
sum(dm$Lineages$rc \%in\% S$Lineages$rc)/nrow(dm$Lineages)  ### at a 53\% FPR
}
