% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_winner.R
\name{filter_winner}
\alias{filter_winner}
\title{filter winner from lineages found}
\usage{
filter_winner(winner, Lineages, rc_table, row.nodemap)
}
\arguments{
\item{winner}{index - element of Lineages maximizing score}

\item{rc_table}{\code{\link{makeRCtable}}}

\item{row.nodemap}{\code{\link{makeNodeMap}}}
}
\description{
filter winner from lineages found
}
