% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lineages.R
\name{find_lineages}
\alias{find_lineages}
\title{find lineages with RCmap}
\usage{
find_lineages(RCmap, rc_table, row.nodemap. = row.nodemap,
  col.nodemap. = col.nodemap)
}
\arguments{
\item{RCmap}{see \code{\link{makeRCmap}}}

\item{rc_table}{see \code{\link{makeRCtable}}}

\item{row.nodemap}{\code{\link{makeNodeMap}} of row.tree}

\item{col.nodemap}{\code{\link{makeNodeMap}} of col.tree}
}
\description{
find lineages with RCmap
}
