% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lineages.R
\name{find_lineages}
\alias{find_lineages}
\title{find lineages with RCmap}
\usage{
find_lineages(RCmap, rc_table, Row_Descendants, Col_Descendants, cl)
}
\arguments{
\item{RCmap}{see \code{\link{makeRCmap}}}

\item{rc_table}{see \code{\link{makeRCtable}}}

\item{Row_Descendants}{named \code{getIndexSets} of all row.nodes}

\item{Col_Descendants}{named \code{getIndexSets} of all col.nodes}

\item{cl}{cluster with library \code{dendromap} loaded on each worker}
}
\description{
find lineages with RCmap
}
