% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRCmap.R
\name{makeRCMap}
\alias{makeRCMap}
\title{make rc-rc mapping for ancestor-descendant relationships & rc tree traversal}
\usage{
makeRCMap(rc_table, row.nodemap, col.nodemap)
}
\arguments{
\item{rc_table}{made from \code{\link{makeRCtable}}}

\item{row.nodemap}{made from \code{\link{makeNodeMap}} of row.tree}

\item{col.nodemap}{made from \code{\link{makeNodeMap}} of col.tree}
}
\description{
make rc-rc mapping for ancestor-descendant relationships & rc tree traversal
}
