% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_seqs.R
\name{rc_seqs}
\alias{rc_seqs}
\title{Walk ant path from descendant - get sequence of ancestral rc's from distal rc}
\usage{
rc_seqs(i, RCmap)
}
\arguments{
\item{i}{index - wihch descendant of \code{RCmap} to for which to find ancestral rc sequence}

\item{RCmap}{made in \code{\link{makeRCmap}}}
}
\description{
Walk ant path from descendant - get sequence of ancestral rc's from distal rc
}
